+++ libsolidity/ast/AST.h
@@ -653,8 +653,9 @@ private:
  * all functions, and this is checked at the point of the using statement. For versions 1 and
  * 2, this check is only done when a function is called.
  *
- * For version 4, T has to be user-defined value type and the function must be pure.
+ * For version 4, T has to be user-defined value type or struct and the function must be pure.
  * All parameters and return value of all the functions have to be of type T.
+ * Locations of all parameters and the return value must be identical.
  * This version can be combined with version 3 - a single directive may attach functions to the
  * type and define operators on it at the same time.
  *