+++ test/libsolidity/syntaxTests/operators/userDefined/defining_operator_for_builtin_types.sol
@@ -6,8 +6,8 @@ using {f as +} for string;
 
 function f(uint, uint) pure returns (uint) {}
 // ----
-// TypeError 5332: (7-8): Operators can only be implemented for user-defined value types.
-// TypeError 5332: (32-33): Operators can only be implemented for user-defined value types.
-// TypeError 5332: (60-61): Operators can only be implemented for user-defined value types.
-// TypeError 5332: (102-103): Operators can only be implemented for user-defined value types.
-// TypeError 5332: (158-159): Operators can only be implemented for user-defined value types.
+// TypeError 5332: (7-8): Operators can only be implemented for user-defined value types and structs.
+// TypeError 5332: (32-33): Operators can only be implemented for user-defined value types and structs.
+// TypeError 5332: (60-61): Operators can only be implemented for user-defined value types and structs.
+// TypeError 5332: (102-103): Operators can only be implemented for user-defined value types and structs.
+// TypeError 5332: (158-159): Operators can only be implemented for user-defined value types and structs.