+++ test/libsolidity/syntaxTests/operators/userDefined/defining_operator_for_contract.sol
@@ -7,5 +7,5 @@ function fa(A, A) pure returns (A) {}
 contract C {}
 abstract contract A {}
 // ----
-// TypeError 5332: (7-9): Operators can only be implemented for user-defined value types.
-// TypeError 5332: (30-32): Operators can only be implemented for user-defined value types.
+// TypeError 5332: (7-9): Operators can only be implemented for user-defined value types and structs.
+// TypeError 5332: (30-32): Operators can only be implemented for user-defined value types and structs.