+++ test/libsolidity/syntaxTests/operators/userDefined/defining_operator_for_interface.sol
@@ -4,4 +4,4 @@ function f(I, I) pure returns (I) {}
 
 interface I {}
 // ----
-// TypeError 5332: (7-8): Operators can only be implemented for user-defined value types.
+// TypeError 5332: (7-8): Operators can only be implemented for user-defined value types and structs.