+++ test/libsolidity/syntaxTests/operators/userDefined/defining_operator_for_struct.sol
@@ -5,5 +5,3 @@ struct S {
 }
 
 function add(S memory, S memory) pure returns (S memory) {}
-// ----
-// TypeError 5332: (7-10): Operators can only be implemented for user-defined value types.