+++ Changelog.md
@@ -2,7 +2,7 @@
 
 Language Features:
 * Allow named parameters in mapping types.
-* Allow defining custom operators for user-defined value types via ``using {f as +} for Typename;`` syntax.
+* Allow defining custom operators for user-defined value types and structs via ``using {f as +} for Typename;`` syntax.
 
 
 Compiler Features: